pyproject.toml
[project]
name = "surkai"
version = "0.1.0"
description = "Frederico Zolio Gonzaga Diniz - fredzolio@live.com"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "qdrant-client>=1.14.3",
    "sentence-transformers>=5.0.0",
    "uvicorn>=0.35.0",
]
